+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+== 1.1.1 (May 24, 2014)
+
+* Fix issues with the toolbar not hiding in certain conditions, even if the selection was lost
+
 == 1.1 (May 23, 2014)
 
 * Fix error thrown in certain cases when the editor was loosing focus